move outputbuffer under io
