+++ src/test/scala/com/github/sadikovi/serde/ConvertersSuite.scala
@@ -27,6 +27,7 @@ import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.UTF8String
 
 import com.github.sadikovi.serde.Converters._
+import com.github.sadikovi.serde.io.OutputBuffer
 import com.github.sadikovi.testutil.UnitTestSuite
 
 class ConvertersSuite extends UnitTestSuite {
